Preliminary Refactoring of graphics painting functions
NOTE: Code does not work at this point.
